requirements.txt
langgraph>=0.2.40
langchain-core>=0.3.0
anthropic>=0.39.0
openai>=1.54.0
streamlit>=1.39.0
pydantic>=2.9.0
python-dotenv>=1.0.1
reportlab>=4.2.5
PyYAML>=6.0.2
langsmith>=0.1.140
pytest>=8.3.0
